requirements.txt
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.4
bs4==0.0.1
certifi==2023.7.22
chardet==3.0.4
Click==7.0
#cloud-init==23.1.2
colorama==0.4.3
#command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contenttype==0.1.3
crispy-forms-gds==0.2.4
cryptography==2.8
cssselect==1.2.0
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
#distro-info===0.23ubuntu1
Django==3.2.8
django-auto-logout==0.5.1
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-js-asset==2.1.0
django-seed==0.3.1
#ec2-hibinit-agent==1.0.0
entrypoints==0.3
executing==2.0.1
fake-useragent==1.3.0
Faker==19.13.0
#hibagent==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==6.1.0
incremental==16.10.1
ipython==8.12.3
jedi==0.19.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.3
MarkupSafe==1.1.0
matplotlib-inline==0.1.6
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
parse==1.19.1
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyee==8.2.2
Pygments==2.16.1
#PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.15.5
pyserial==3.4
#python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3.post1
PyYAML==5.3.1
requests==2.22.0
requests-html==0.10.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
toposort==1.10
tqdm==4.66.1
traitlets==5.13.0
Twisted==18.9.0
typing-extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uWSGI==2.0.23
w3lib==2.1.2
wadllib==1.3.3
wcwidth==0.2.9
websockets==10.4
zipp==3.17.0
zope.interface==4.7.1
